✨ feat: PortfolioTrainings Entity 작성
