+++ src/main/java/com/cheeup/domain/PortfolioTrainings.java
@@ -1,0 +1,27 @@
+package com.cheeup.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.Date;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class PortfolioTrainings {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 255)
+    private String name;
+
+    @Column(nullable = false)
+    private Date startDate;
+
+    private Date endDate;
+
+    private String content;
+}
