+++ src/main/java/com/cheeup/domain/PostLikes.java
@@ -1,4 +1,4 @@
-package com.cheeup.domain.mapping;
+package com.cheeup.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -13,7 +13,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
-public class FavoriteBoards {
+public class PostLikes {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
